app/cloud-sql-query-runner/requirements.txt
#google-api-python-client==1.7.11
google-cloud-pubsub==1.1.0
SQLAlchemy==1.3.13
PyMySQL==0.9.3
